Add method, release, and compile tests.
